 requirements.txt
# Core libraries
numpy==1.21.0
pandas==1.3.0

# Data ingestion and API requests
requests==2.26.0
google-cloud-storage==1.42.0

# Database connectors
pymysql==1.0.2  # For MySQL, as it's used in the ingestion script

# Data encryption and security for LGPD compliance
cryptography==3.4.7

# Scheduler for automated jobs (if needed)
schedule==1.1.0
